Hide package-tab; Combine multiple features in single file
